Move GameOptionsModel to common.
