+++ src/client/components/pathfinders/PlanetaryTrackReward.vue
@@ -9,7 +9,7 @@
 <script lang="ts">
 
 import Vue from 'vue';
-import {GameOptionsModel} from '@/models/GameOptionsModel';
+import {GameOptionsModel} from '@/common/models/GameOptionsModel';
 import {Reward} from '@/common/pathfinders/Reward';
 
 export default Vue.extend({
