+++ src/models/ServerModel.ts
@@ -3,7 +3,7 @@ import {ColonyModel} from '../common/models/ColonyModel';
 import {Color} from '../common/Color';
 import {Game, GameOptions} from '../Game';
 import {SimpleGameModel} from './SimpleGameModel';
-import {GameOptionsModel} from './GameOptionsModel';
+import {GameOptionsModel} from '../common/models/GameOptionsModel';
 import {ICard} from '../cards/ICard';
 import {IProjectCard} from '../cards/IProjectCard';
 import {isICloneTagCard} from '../cards/pathfinders/ICloneTagCard';
